Parse resource name annotations from imported protos too (#2998)
So far we only parse resource name related annotations from source protos (proto files that we generate code for) but not imports. However this does not work for IAM. IAM request messages declare resource_reference.type = "*", but the request object is an import instead of a source proto for a certain API.

Also slightly tweak the method addTag (add method_signature and change resource_reference.type to "*") to mimic IAM methods.