Support WORKSPACE-relative paths as command line arguments for the bazel biuld file generator. (#3057)
Also make `--dest` argument optional (by default it will be equal to `--src`, meaning that the generator will overwrite `BUILD.bazel` file in the `--src` directory (most common scenario)).

The `BUILD_WORKSPACE_DIRECTORY` environment variable is set automatically by bazel for all targets ran as `bazel run`. The value of the env varible is the path to the root folder of the repository.