fix: honor resource config on name field in method flattening (#3115)
We generate flattening overload methods where resource name strings are replaced by resource name types. Previously we only check resource_reference to see if a field is a resource name, this is not enough as the field `name` in a resource message is also a resource name field, but does not have resource_reference annotation.

I found this while generating the method CreateSubscription for PubSub. The pattern is strongly discouraged for new APIs; but we can't break PubSub.