Assign the proto file where the interface is defined when generating samples and tests for arbitrary resource names(#3138)
Follow-up of #3027

When picking a SingleResourceNameConfig to use to generate samples for AnyResourceNameConfig, we need to assign the correct proto file to the SingleResourceNameConfig so that we know the correct full classpath of the generated resource name class.

#3027 uses the package where the method is defined, which fixed a common case of IAM methods where request objects are imported from the IAM package.

That did not cover the case for PubSub where the IAM methods are added by reroute_to_grpc_interface in gapic config. In this case the method is also defined in IAM package.

This PR change the behavior to assigning the proto file where the interface we are generating is defined, which should cover both cases.