chore: reorganize tests (#3141)
gapic-generator has too many features sharing the same set of testing source files and it's been harder and harder to keep the baseline changes to a minimum when testing a new feature.

This PR restructures the testing sources as follows:

- gapic config clients + gapic config samples (the old java_library.baseline):
  - deleted as we no longer need samples configured by gapic YAMLs
- gapic config clients + sample config samples (the old 
 sample_config_migration_java_library.baseline):
  - use test sources in testsrc/gapicconfig directory
  - the baseline file is renamed to java_library.baseline
- proto annotations clients (including some unit tests):
  - use test sources in testsrc/protoannotations directory
- packagegen:
  - continue to use test sources in testsrc/common
- configgen:
  - continue to use test sources in testsrc/common

Going forward there should be minimal changes to testsrc/common and testsrc/gapicconfig as well as respective baseline files as these functionality are stable. This would make working on new features/bug fixes to proto annotation based generation easier.