all: set max & initial timeouts to total timeout
When using grpc_service_config the max & initial rpc timeout
should be the same as the total timeout. This essentially
renders the hedging logic unused because the initial timeout
is the limit. This means one RPC is made rather than several
smaller ones that are backed off until the total is reached.

* switch showcase to gapic v2 & grpc_service_config
* retry mapping: fix NPE when a Service is not present in protos
* add blockTimeout showcase test
* add valid block showcase test
* update gapic-showcase image for java to v0.10.1
* lengthen delay on flaky showcase java streaming tests