fix: use separate value generator in test generation (#3294)
Sometimes a library has two requests methods with a field of the same name but different types. The TestValueGenerator keys on the short name of fields, and reuses a generated value if it has generated a value for this key (short field name). Therefore in such cases it may try to assign a string literal to a integer type.

Change to create a different TestValueGenerator for every single test case generation to avoid such cases.